pyproject.toml
[project]
name = "langchain-e-python-criando-ferramentas-com-a-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-experimental>=0.3.4",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.7",
    "openai>=1.107.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
]
